requirements.txt
click==6.7
Flask==1.0.2
gevent==1.3.6
greenlet==0.4.15
gunicorn==19.9.0
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
PyYAML==3.13
Werkzeug==0.14.1
